fix(saml): separate view from helper to avoid circular deps
